ECT support, fixes #2187
git-svn-id: https://subversion.assembla.com/svn/kinvey/ios-library/trunk@6175 868178e2-d38b-4ba8-a603-825f91c87007

+++ KinveyKit/KinveyKit/Source/KCSLinkedAppdataStore.m
@@ -28,7 +28,7 @@
 {
     //Step 2: Save References
     NSArray* references = so.referencesToSave;
-    int totalReferences = references.count;
+    NSUInteger totalReferences = references.count;
     
     if (totalReferences == 0) {
         //no references, go on to saving object
@@ -94,7 +94,7 @@
 {
     //Step 1: Save Resources
     NSArray* resources = so.resourcesToSave;
-    int totalResources = resources.count;
+    NSUInteger totalResources = resources.count;
     
     if (totalResources == 0) {
         //no resources, go on to saving references
